app/pyproject.toml
[tool.poetry]
package-mode=false
name = "auto_index"
version = "1.0.0"
description = ""
authors = ["Viet Do <vietdqhcmute@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.95.1"
python-docx = "^1.1.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.2.0"
ruff = "^0.0.263"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
